balancedbrief/app/requirements.txt
openai
psycopg2-binary
bs4
newspaper3k
datetime
praw
premailer
boto3
cssutils
slack_sdk
argparse
pytz